working on new sub box image uploading issue

+++ app/Http/Controllers/ContentController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Models\{Field,FormStructure,SubformStructure, Category, ApplicationData, AppData, SubAppData};
 use Yajra\DataTables\DataTables;
 use Illuminate\Support\Str;
+use Carbon\Carbon;
 
 class ContentController extends Controller
 {
@@ -114,24 +115,24 @@ class ContentController extends Controller
         $application_id = $id;
 
         // dump($data);
-        $main_structure = FormStructure::where('application_id', $application_id)->get()->pluck('id')->toArray();
-        $sub_structure = SubformStructure::where('application_id', $application_id)->get()->pluck('id')->toArray();
-        $sub_structure_formid = SubformStructure::where('application_id', $application_id)->first();
-        // dd($sub_structure_formid);
-        if($sub_structure_formid != null && !empty($sub_structure_formid)){
-            $sub_form_id = $sub_structure_formid->form_id;
-        }
         $field_names = (isset($data['field_name']) && $data['field_name']) ? $data['field_name'] : null;
         $field_type = (isset($data['field_type']) && $data['field_type']) ? $data['field_type'] : null;
         $sub_field_names = (isset($data['sub_field_name']) && $data['sub_field_name']) ? $data['sub_field_name'] : null;
         $sub_field_type = (isset($data['sub_field_type']) && $data['sub_field_type']) ? $data['sub_field_type'] : null;
         // $UUID_main = (isset($data['UUID-main']) && $data['UUID-main']) ? $data['UUID-main'] : null;
+        $main_structure = FormStructure::where('application_id', $application_id)->get()->pluck('id')->toArray();
+        $sub_structure = SubformStructure::where('application_id', $application_id)->get()->pluck('id')->toArray();
+        $sub_structure_formid = SubformStructure::where('application_id', $application_id)->first();
+        if($sub_structure_formid != null && !empty($sub_structure_formid)){
+            $sub_form_id = $sub_structure_formid->form_id;
+        }
+        $get_all_content = AppData::where('app_id', $application_id)->groupBy('UUID')->get();
+        $get_all_content_if = AppData::where('app_id', $application_id)->first();
+        $get_all_sub_content = SubAppData::where('app_id', $application_id)->groupBy('UUID')->get();
+        $get_all_sub_content_if = SubAppData::where('app_id', $application_id)->first();
+        
+        // dd($get_all_content);
 
-        // dump($UUID_main);
-        // dump($field_names);
-        // dump($field_type);
-        // dump($sub_field_names);
-        // dd($sub_field_names);
         $from_struucture_array = [];
         $sub_from_struucture_array = [];
         if($field_names != ""){
@@ -143,7 +144,20 @@ class ContentController extends Controller
                 $FormStructures->created_by = \Auth::id();
                 $FormStructures->save();
 
+                if($get_all_content != null && !empty($get_all_content)){
+                    foreach($get_all_content as $content){
+                        $app_data = new AppData();
+                        $app_data->UUID = $content->UUID;
+                        $app_data->app_id = $content->app_id;
+                        $app_data->category_id = $content->category_id;
+                        $app_data->form_structure_id = $FormStructures->id;
+                        $app_data->value = " ";
+                        $app_data->save();
+                    }
+                }
+
                 array_push($from_struucture_array, $FormStructures->id);
+                // dump($field_type[$key]);
                 if($field_type[$key] == "sub-form"){
                     foreach($sub_field_names as $subkey => $sub_field_name){
                         $SubFormStructures = new SubformStructure();
@@ -153,22 +167,56 @@ class ContentController extends Controller
                         $SubFormStructures->field_type = $request->sub_field_type[$subkey];
                         $SubFormStructures->created_by = \Auth::id();
                         $SubFormStructures->save();
+
+                        // dump($SubFormStructures->id);
+
+                        if($get_all_sub_content_if != null && !empty($get_all_sub_content_if)){
+                            // dump($get_all_sub_content);
+                            foreach($get_all_sub_content as $sub_content){
+                                $sub_app_data = new SubAppData();
+                                $sub_app_data->app_id = $sub_content->application_id;
+                                $sub_app_data->category_id = $sub_content->category_id;
+                                $sub_app_data->app_uuid = $sub_content->app_uuid;
+                                $sub_app_data->UUID = $sub_content->UUID;
+                                $sub_app_data->sub_form_structure_id = $SubFormStructures->id;
+                                $sub_app_data->value = "";
+                                $sub_app_data->save();
+                            }
+                        }else{
+                            if($get_all_content_if != null && !empty($get_all_content_if)){
+                                $ddd = 0;
+                                foreach($get_all_content as $content){
+                                    $current_timestamp = Carbon::now()->timestamp;
+                                    $sub_app_data = new SubAppData();
+                                    $sub_app_data->app_id = $content->app_id;
+                                    $sub_app_data->category_id = $content->category_id;
+                                    $sub_app_data->app_uuid = $content->UUID;
+                                    $sub_app_data->UUID = $current_timestamp."_s".$ddd;
+                                    $sub_app_data->sub_form_structure_id = $SubFormStructures->id;
+                                    $sub_app_data->value = "";
+                                    $sub_app_data->save();
+                                    $ddd ++;
+                                }
+                            }
+                        }
                     }
                 }
-                // else{
             }
+            // dd();
         }else{
-            if($sub_field_type != null){
-                foreach($sub_field_names as $subkey => $sub_field_name){
-                    $SubFormStructures = new SubformStructure();
-                    $SubFormStructures->application_id = $application_id;
-                    $SubFormStructures->form_id = $sub_form_id;
-                    $SubFormStructures->field_name = $sub_field_name;
-                    $SubFormStructures->field_type = $sub_field_type[$subkey];
-                    $SubFormStructures->created_by = \Auth::id();
-                    $SubFormStructures->save();
+            if($sub_structure_formid != null && !empty($sub_structure_formid)){
+                if($sub_field_type != null){
+                    foreach($sub_field_names as $subkey => $sub_field_name){
+                        $SubFormStructures = new SubformStructure();
+                        $SubFormStructures->application_id = $application_id;
+                        $SubFormStructures->form_id = $sub_form_id;
+                        $SubFormStructures->field_name = $sub_field_name;
+                        $SubFormStructures->field_type = $sub_field_type[$subkey];
+                        $SubFormStructures->created_by = \Auth::id();
+                        $SubFormStructures->save();
 
-                    array_push($sub_from_struucture_array, $SubFormStructures->id);
+                        array_push($sub_from_struucture_array, $SubFormStructures->id);
+                    }
                 }
             }
         }
@@ -214,46 +262,48 @@ class ContentController extends Controller
                                             ->delete();
             $subform_remove_form_fields->delete();
         }else{
-            $sub_form_fieldss = SubformStructure::where('application_id', $application_id)->first();
-            if($sub_form_fieldss != null){
-                $other_form_fields1 = FormStructure::where('application_id', $application_id)
-                                    ->whereNotIn('id', $from_struucture_array)
-                                    ->get()->pluck('id')->toArray();
-                $main_data_delete = AppData::whereIn('form_structure_id', $other_form_fields1)->where('app_id', $application_id)->delete();
+            if($sub_structure_formid != null && !empty($sub_structure_formid)){
+                $sub_form_fieldss = SubformStructure::where('application_id', $application_id)->first();
+                if($sub_form_fieldss != null){
+                    $other_form_fields1 = FormStructure::where('application_id', $application_id)
+                                        ->whereNotIn('id', $from_struucture_array)
+                                        ->get()->pluck('id')->toArray();
+                    $main_data_delete = AppData::whereIn('form_structure_id', $other_form_fields1)->where('app_id', $application_id)->delete();
 
-                $form_fields = FormStructure::where('application_id', $application_id)
-                                                ->whereNotIn('id', $from_struucture_array)
-                                                ->delete();
+                    $form_fields = FormStructure::where('application_id', $application_id)
+                                                    ->whereNotIn('id', $from_struucture_array)
+                                                    ->delete();
 
-                $other_form_fields2 = SubformStructure::where('application_id', $application_id)
-                        ->whereNotIn('id', $from_struucture_array)
-                        ->where('form_id', $sub_form_id)
-                        ->get()->pluck('id')->toArray();
-                $main_data_delete1 = SubAppData::where('app_id', $application_id)
-                                    ->whereIn('sub_form_structure_id', $other_form_fields2)
-                                    ->delete();
-                $sub_form_fields = SubformStructure::where('application_id', $application_id)
-                                                ->where('form_id', $sub_form_id)
-                                                ->whereNotIn('id', $sub_from_struucture_array)
-                                                ->delete();
-                foreach($data as $key => $dd){
-                    $splitd = explode("_",$key);
-                    $rocord_id = $splitd[0];
-                    $type = $splitd[1];
-                    $match = $splitd[2];
-                    
-                    $form_fieldsa = FormStructure::where('application_id', $application_id)->where('id', $rocord_id)->first();
-                    $sub_form_fieldsa = SubformStructure::where('application_id', $application_id)->where('id', $rocord_id)->first();
-        
-                    if($match == "name"){
-                        if($type == "main"){
-                            $form_fieldsa->field_name = $dd[0];
-                            $form_fieldsa->save();
-                            array_push($from_struucture_array, $rocord_id);
-                        }else if($type == "sub"){
-                            array_push($sub_from_struucture_array, $rocord_id);
-                            $sub_form_fieldsa->field_name = $dd[0];
-                            $sub_form_fieldsa->save();
+                    $other_form_fields2 = SubformStructure::where('application_id', $application_id)
+                            ->whereNotIn('id', $from_struucture_array)
+                            ->where('form_id', $sub_form_id)
+                            ->get()->pluck('id')->toArray();
+                    $main_data_delete1 = SubAppData::where('app_id', $application_id)
+                                        ->whereIn('sub_form_structure_id', $other_form_fields2)
+                                        ->delete();
+                    $sub_form_fields = SubformStructure::where('application_id', $application_id)
+                                                    ->where('form_id', $sub_form_id)
+                                                    ->whereNotIn('id', $sub_from_struucture_array)
+                                                    ->delete();
+                    foreach($data as $key => $dd){
+                        $splitd = explode("_",$key);
+                        $rocord_id = $splitd[0];
+                        $type = $splitd[1];
+                        $match = $splitd[2];
+                        
+                        $form_fieldsa = FormStructure::where('application_id', $application_id)->where('id', $rocord_id)->first();
+                        $sub_form_fieldsa = SubformStructure::where('application_id', $application_id)->where('id', $rocord_id)->first();
+            
+                        if($match == "name"){
+                            if($type == "main"){
+                                $form_fieldsa->field_name = $dd[0];
+                                $form_fieldsa->save();
+                                array_push($from_struucture_array, $rocord_id);
+                            }else if($type == "sub"){
+                                array_push($sub_from_struucture_array, $rocord_id);
+                                $sub_form_fieldsa->field_name = $dd[0];
+                                $sub_form_fieldsa->save();
+                            }
                         }
                     }
                 }
